Restructure eclipse project to maven project
