+++ src/main/java/io/github/danielahaeussler/heldbuilder/game/LevelSystem.java
@@ -1,4 +1,4 @@
-package game;
+package io.github.danielahaeussler.heldbuilder.game;
 
 public class LevelSystem {
 	int[] level;
