+++ src/main/java/io/github/danielahaeussler/heldbuilder/helden/EKlassen.java
@@ -1,4 +1,4 @@
-package helden;
+package io.github.danielahaeussler.heldbuilder.helden;
 
 public enum EKlassen {
