+++ src/main/java/io/github/danielahaeussler/heldbuilder/helden/ESpecies.java
